+++ Backend/src/main/java/mk/awd/informirajse/model/Blog.java
