+++ Backend/src/main/java/mk/awd/informirajse/model/Comment.java
